Data from previous National Park project
This is data from Jason's previous national parks project. We could probably get more variables if we need, but this was our final analysis dataset. The precipitation data is also included, and can be used if we subset the national parks data later. It is not really usable with the full database.